# Speed up tests by subsampling