--- a/docs/installation.md
+++ b/docs/installation.md
@@ -71,11 +71,14 @@ Follow these steps to install pertpy on an Apple Silicon machine (tested on a Ma
     ```
 
 5. Go inside the pertpy folder and install pertpy
+
     ```console
     $ cd pertpy
     $ pip install .
     ```
+
     Now you're ready to use pertpy as usual within the environment (`import pertpy`).
+
     ```
 
     ```