--- a/tests/tools/_coda/test_sccoda.py
+++ b/tests/tools/_coda/test_sccoda.py
@@ -3,6 +3,7 @@ from pathlib import Path
 import arviz as az
 import numpy as np
 import pandas as pd
+import scanpy as sc
 import pytest
 from mudata import MuData
 
@@ -22,6 +23,8 @@ class TestscCODA:
     @pytest.fixture
     def adata(self):
         cells = pt.dt.haber_2017_regions()
+        cells = sc.pp.subsample(cells, 0.1, copy=True)
+
         return cells
 
     def test_load(self, adata):