--- a/tests/tools/test_cinemaot.py
+++ b/tests/tools/test_cinemaot.py
@@ -12,6 +12,7 @@ class TestCinemaot:
     @fixture
     def adata(self):
         adata = pt.dt.cinemaot_example()
+        adata = sc.pp.subsample(adata, 0.1, copy=True)
 
         return adata
 
@@ -74,8 +75,9 @@ class TestCinemaot:
             adata, de, pert_key="perturbation", control="No stimulation", label_list=None
         )
 
-        expect_num = 20
+        expect_num = 9
         eps = 7
         assert "ptb" in adata_pb.obs
         assert not np.isnan(np.sum(adata_pb.X))
+        print(adata_pb.shape)
         assert not np.abs(adata_pb.shape[0] - expect_num) >= eps