<!-- Many thanks for contributing to this project! -->

**PR Checklist**

<!-- Please fill in the appropriate checklist below (delete whatever is not relevant). These are the most common things requested on pull requests (PRs). -->

-   [X] Referenced issue is linked (See #370)

**Description of changes**

In order to speed up testing, the data used for testing can be subsampled. Here, I examined all tests that use a pertpy dataloader and performed subsampling on the corresponding `AnnData` object.

To ensure that all tests still run, I had to make modifications to two elements that were hardcoded based on the size of the respective data object:
- In `test_tasccoda.test_prepare`: `assert np.sum(mdata["coda"].obsm["covariate_matrix"]) == 8` instead of `== 85`
- In `test_cinemaot.test_pseudobulk`: `expect_num = 9` instead of `expect_num = 20` (because `adata_pb.shape[0]` is 9 when subsampled by 10%)


_Note:_ `TestAugur`, which also relies on pertpy data, has not been modified yet. The tests will be rewritten as described in issue #380, and the subsampling will then be incorporated into the PR for that issue.


**Technical details**

I have confirmed that all tests still run successfully after subsampling. The following improvements in test running time were achieved (tested on my personal computer):

`test_sccoda`
- No subsampling: 43 sec 289 ms
- Subsampled by 10%: 38 sec 752 ms
- Subsampling by 5% or less resulted in an error in `test_credible_effects`, so I kept 10%

`test_tasccoda`
- No subsampling: 35 sec 578 ms
- Subsampled by 10%: 8 sec 828 ms

`test_cinemaot`
- No subsampling: 1 min 9 sec
- Subsampled by 10%: 8 sec 734 ms